Removed celery
