+++ backend/recipes/__init__.py
@@ -152,7 +152,6 @@ def stop(_):
     """
     Stop the currently running recipe.
 
-    TODO: Implement mechanism for stopping associated celery tasks
     :return:
     None ... at least for now.
     """
