+++ backend/requirements.txt
@@ -1,8 +1,6 @@
 Flask==2.3.2
 flask-cors==3.0.9
 pytest==5.4.1
-celery==5.2.2
-redis==3.4.1
 requests==2.31.0
 pyserial==3.5
 w1thermsensor==2.0.0
