--- scripts/startCelery.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-cd ./backend
-virtualenv -p python3 env
-source env/bin/activate
-pip install -r requirements.txt
-
-celery -A recipes worker --loglevel=INFO
